refactoring : MainSlide 컴퍼넌트 하나의 너무 많은 기능과 책임이 있는거 같아서 책임을 분리하고 기능을 분리완료
- 메인 이미지슬라이드 컴퍼넌트로 분리
- 텍스트 이미지슬라이드 컴퍼넌트로 분리
- 메인 슬라이드 카운트하는 컴퍼넌트로 분리
- 메인 슬라이드 progress하는 컴퍼넌트로 분리
- 메인 슬라이드 컨트롤하는 컴퍼넌트로 분리
- 비지니스 로직, ui을 각각 분리
- 이미지 컨트롤 제거하고 FontAwesomeIcon으로 변경
